+++ Source/Game/Entities/Player/Player.Achievement.cs
@@ -52,7 +52,7 @@ namespace Game.Entities
             m_questObjectiveCriteriaMgr.ResetCriteria(failEvent, failAsset, evenIfCriteriaComplete);
         }
 
-        public void UpdateCriteria(CriteriaType type, float miscValue1, float miscValue2 = 0, float miscValue3 = 0, WorldObject refe = null)
+        public void UpdateCriteria(CriteriaType type, double miscValue1, double miscValue2 = 0, double miscValue3 = 0, WorldObject refe = null)
         {
             UpdateCriteria(type, (ulong)miscValue1, (ulong)miscValue2, (ulong)miscValue3, refe);
         }
